[dagit] Move repo locations list into Workspace overview
